pyproject.toml
[project]
name = "url-citation-classification-intermediate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.12.0",
    "hydra-core>=1.3.2",
    "nltk>=3.9.3",
    "numpy>=2.3.5",
    "pandas>=3.0.1",
    "scikit-learn>=1.8.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.2.0",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu128" }]
torchvision = [{ index = "pytorch-cu128" }]
